split and fix examples
